+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 - Implement `xsd:unsignedShort` data type codec
 - Implement `xsd:positiveInteger` data type codec
 - Implement `xsd:nonPositiveInteger` data type codec
+- Implement `xsd:unsignedLong` data type codec